Seuls les sites actifs peuvent être associés à des véhicules ou des emprunts + partial_update pour switch etat + validation relations avant suppression

+++ backend/views/view_site.py
@@ -5,6 +5,8 @@ from rest_framework.decorators import permission_classes, action
 from rest_framework.settings import api_settings
 from rest_framework.response import Response
 from backend.models.model_site import Site
+from backend.models.model_emprunt import Emprunt
+from backend.models.model_vehicule import Vehicule
 from backend.serializers.serializer_site import SiteSerializer
 from backend.permissions.permission_site import SitePermission
 
@@ -21,6 +23,17 @@ class SiteViewSet(viewsets.ViewSet):
     
     def list(self, request):
         queryset = Site.objects.all()
+        serializer = SiteSerializer(queryset, many=True)
+        return Response(data= { 'IsSuccess': True, 'Data': serializer.data }, status=status.HTTP_200_OK)
+
+    @action(detail=False, url_path='availables', url_name='list_availables')
+    def list_availables(self, request):
+        """
+        On récupère la liste de tous les sites non archivés
+        """
+        params = request.query_params
+        queryset = Site.objects.exclude(is_active=False).order_by('id')
+
         serializer = SiteSerializer(queryset, many=True)
         return Response(data= { 'IsSuccess': True, 'Data': serializer.data }, status=status.HTTP_200_OK)
 
@@ -50,9 +63,21 @@ class SiteViewSet(viewsets.ViewSet):
         serializer.save()
         return Response(data= { 'IsSuccess': True, 'Data': serializer.data }, status=status.HTTP_200_OK)
 
+    def partial_update(self, request, *args, **kwargs):
+        kwargs['partial'] = True
+        # TODO : Maybe some tests for exceptions here ? like a 404
+        return self.update(request, *args, **kwargs)
+
     def destroy(self, request, pk=None, *args, **kwargs):
         queryset = Site.objects.all()
         site = get_object_or_404(queryset, pk=pk)
+        
+        if Vehicule.objects.filter(site_id=site.id).exists():
+            return Response(data= { 'IsSuccess': False, 'LibErreur' : "Impossible de supprimer ce site car il est lié à un véhicule (privilégier l'archivage)"}, status=status.HTTP_200_OK)
+
+        if Emprunt.objects.filter(site_id=site.id).exists():
+            return Response(data= { 'IsSuccess': False, 'LibErreur' : "Impossible de supprimer ce site car il est lié à un emprunt (privilégier l'archivage)"}, status=status.HTTP_200_OK)
+
         site.delete()
         return Response(data= { 'IsSuccess': True, 'Data': True }, status=status.HTTP_200_OK)
